requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==3.2.3
django-apscheduler==0.6.2
django-filter==23.3
mysqlclient==2.2.0
Pillow==10.0.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
